+++ src/OctoshiftCLI.Tests/ado2gh/Commands/AddTeamToRepoCommandTests.cs
@@ -32,7 +32,7 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
             var team = "foo-team";
             var role = "maintain";
 
-            var mockGithub = new Mock<GithubApi>(null);
+            var mockGithub = new Mock<GithubApi>(null, null);
             var mockGithubApiFactory = new Mock<GithubApiFactory>(null, null, null);
             mockGithubApiFactory.Setup(m => m.Create()).Returns(mockGithub.Object);
 
@@ -50,7 +50,7 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
             var team = "foo-team";
             var role = "read";  // read is not a valid role
 
-            var mockGithub = new Mock<GithubApi>(null);
+            var mockGithub = new Mock<GithubApi>(null, null);
             var mockGithubApiFactory = new Mock<GithubApiFactory>(null, null, null);
             mockGithubApiFactory.Setup(m => m.Create()).Returns(mockGithub.Object);