+++ src/gei/GithubApiFactory.cs
@@ -19,21 +19,36 @@ namespace OctoshiftCLI.GithubEnterpriseImporter
         {
             var githubPat = _environmentVariableProvider.SourceGithubPersonalAccessToken();
             var githubClient = new GithubClient(_octoLogger, _clientFactory.CreateClient("Default"), githubPat);
-            return new GithubApi(githubClient);
+
+            return new GithubApi(githubClient, "https://api.github.com");
         }
 
-        GithubApi ISourceGithubApiFactory.CreateClientNoSSL()
+        GithubApi ISourceGithubApiFactory.Create(string apiUrl)
+        {
+            var githubPat = _environmentVariableProvider.SourceGithubPersonalAccessToken();
+            var githubClient = new GithubClient(_octoLogger, _clientFactory.CreateClient("Default"), githubPat);
+            return new GithubApi(githubClient, apiUrl);
+        }
+
+        GithubApi ISourceGithubApiFactory.CreateClientNoSsl(string apiUrl)
         {
             var githubPat = _environmentVariableProvider.SourceGithubPersonalAccessToken();
             var githubClient = new GithubClient(_octoLogger, _clientFactory.CreateClient("NoSSL"), githubPat);
-            return new GithubApi(githubClient);
+            return new GithubApi(githubClient, apiUrl);
         }
 
         GithubApi ITargetGithubApiFactory.Create()
         {
             var githubPat = _environmentVariableProvider.TargetGithubPersonalAccessToken();
             var githubClient = new GithubClient(_octoLogger, _clientFactory.CreateClient("Default"), githubPat);
-            return new GithubApi(githubClient);
+            return new GithubApi(githubClient, "https://api.github.com");
+        }
+
+        GithubApi ITargetGithubApiFactory.Create(string apiUrl)
+        {
+            var githubPat = _environmentVariableProvider.TargetGithubPersonalAccessToken();
+            var githubClient = new GithubClient(_octoLogger, _clientFactory.CreateClient("Default"), githubPat);
+            return new GithubApi(githubClient, apiUrl);
         }
     }
 }