+++ RELEASENOTES.md
@@ -1,1 +1,3 @@
-
+- Rename `gh gei generate-archive` to `gh gei migrate-archive-repo` which generates migration archives, uploads them to azure blob storage, and kicks off an github enterprise importer migration
+- Modify `gh gei migrate-repo` to optionally accept two archive urls to start a migration
+- Add target api url parameter to `gh gei migrate-repo` command